Adds missing .gap-s class
I just noticed that the `.gap-s` class was missing among the gaps while the custom property exists and the corresponding docs page claims that there is a `.gap-s` class.

btw, thanks for this lib! Will try it out soon in a project. Had something like this in mind for a little while ...